Create HashingAlgorithm instances for all four BagIt-supported algos
